Rename ADSC format into d*TREK format
