+++ fabio/test/testfilenames.py
@@ -62,10 +62,10 @@ CASES = [
     (670005, 'edf', 'S82P670005.edf'),
     (670005, 'edf', 'S82P670005.edf.gz'),
     # based on only the name it can be either img or oxd
-    (1, 'adsc_or_oxd_or_hipic_or_raxis', 'mb_LP_1_001.img'),
-    (2, 'adsc_or_oxd_or_hipic_or_raxis', 'mb_LP_1_002.img.gz'),
-    (3, 'adsc_or_oxd_or_hipic_or_raxis', 'mb_LP_1_003.img.bz2'),
-    (3, 'adsc_or_oxd_or_hipic_or_raxis', os.path.join("data", 'mb_LP_1_003.img.bz2')),
+    (1, 'dtrek_or_oxd_or_hipic_or_raxis', 'mb_LP_1_001.img'),
+    (2, 'dtrek_or_oxd_or_hipic_or_raxis', 'mb_LP_1_002.img.gz'),
+    (3, 'dtrek_or_oxd_or_hipic_or_raxis', 'mb_LP_1_003.img.bz2'),
+    (3, 'dtrek_or_oxd_or_hipic_or_raxis', os.path.join("data", 'mb_LP_1_003.img.bz2')),
 ]
